fix: `--help` usage info
yargs (via nconf) would exit when detecting a help flag

also improves the speed of `./nodebb help build`
