+++ install/package.json
@@ -146,6 +146,7 @@
         "winston": "3.3.3",
         "xml": "^1.0.1",
         "xregexp": "^4.3.0",
+        "yargs": "16.2.0",
         "zxcvbn": "^4.4.2"
     },
     "devDependencies": {
